pyproject.toml
[tool.poetry]
name = "mongodb-tut"
version = "0.1.0"
description = "mongodb tutorial"
authors = ["fouad-alef <fouad.shammary@alefeducation.com>"]
readme = "README.md"
packages = [{include = "mongodb_tut"}]

[tool.poetry.dependencies]
python = "^3.9"
pymongo = "^4.3.3"
python-dotenv = "^1.0.0"
beanie = "^1.17.0"
domain = "^0.1.4"
isort = "^5.12.0"
flake8 = "^6.0.0"
black = "^23.1.0"
pydantic = "^1.10.6"
asyncio = "^3.4.3"
motor = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
